requirements.txt
flask==2.3.3
flask-cors==4.0.0
langchain==0.1.0
langchain-community==0.0.10
langchain-google-genai==0.0.6
langchain-groq==0.0.1
chromadb==0.4.22
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
tiktoken==0.5.2
sentence-transformers==2.2.2
langchain-core==0.1.23
# sqlite3 is included in Python stdlib, no need to add 